requirements.txt
absl-py==0.5.0
aniso8601==3.0.2
astor==0.7.1
audioread==2.1.6
awscli==1.14.36
boto==2.49.0
boto3==1.9.26
botocore==1.12.26
bz2file==0.98
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.7
cycler==0.10.0
Cython==0.29
decorator==4.3.0
docutils==0.14
Flask==1.0.2
Flask-RESTful==0.3.6
gast==0.2.0
gensim==3.6.0
grpcio==1.15.0
h5py==2.8.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.5
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
llvmlite==0.25.0
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.1
nltk==3.3
numba==0.40.1
numpy==1.15.2
Pillow==5.3.0
protobuf==3.6.1
pyasn1==0.4.2
pycparser==2.19
pymssql==2.1.4
pyodbc==4.0.24
pyparsing==2.3.0
python-dateutil==2.6.1
python-dotenv==0.9.1
pytz==2018.5
PyYAML==3.12
requests==2.20.0
resampy==0.2.1
rsa==3.4.2
s3transfer==0.1.12
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
smart-open==1.7.1
SoundFile==0.10.2
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
urllib3==1.24
Werkzeug==0.14.1
